pyproject.toml
[project]
name = "assignment-2-aravind-jayachandran-ushakumari"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "seaborn>=0.13.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "google-cloud-storage>=2.11.0",
    "pydantic>=2.9.0" 
    
]
